gsfl-dashboard/requirements.txt
alabaster==0.7.13
astroid==2.15.6
asyncio==3.4.3
attrs==23.1.0
azure-core==1.29.3
azure-storage-blob==12.17.0
blinker==1.6.2
certifi==2023.7.22
cffi==1.15.1
chardet==5.2.0
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
contextlib2==21.6.0
cryptography==3.1
evdev==1.6.1
Flask==2.3.3
idna==3.4
importlib-metadata==6.8.0
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.0.3
json2html==1.3.0
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
lazy-object-proxy==1.9.0
MarkupSafe==2.1.3
mccabe==0.7.0
PGPy==0.5.4
pika==1.3.2
pockets==0.9.1
pyasn1==0.5.0
pycparser==2.21
Pygments==2.16.1
PyJWT==1.6.1
pynput==1.7.6
pyparsing==2.4.7
pyrql==0.7.0
python-benedict==0.32.0
python-dateutil==2.8.2
python-dotenv==0.21.1
python-fsutil==0.10.0
python-gnupg==0.5.1
python-slugify==8.0.1
python-xlib==0.33
referencing==0.30.2
requests==2.31.0
rpds-py==0.10.0
s3i==0.6.5
schema==0.7.5
six==1.16.0
snowballstemmer==2.2.0
text-unidecode==1.3
typed-ast==1.5.5
typing_extensions==4.7.1
urllib3==2.0.4
uuid==1.30
websocket-client==1.6.2
Werkzeug==2.3.7
wrapt==1.15.0
zipp==3.16.2
